fix array offset error on news_item page
Previously, if you've opened the news_item page of a news item, which isn't
the first in the database table, then there will show up a message like
'Fatal error: array offset...'.

The problem was in that the built-in php array_filter() function
always saves keys, and because the way of getting the single news
object from the array was getting it by the key '0', there have been
error raising, because the second, third etc. items were having
'1', '2' etc. keys.

Fix by making a new array with the array_values() function, which helps
reseting the keys in the increasing order, starting from 0.