add small changes
Fix users logging in error (which was due to wrong array index offset,
as in some previous commit was for the news item, and fixed by the same
array_values() function).
Add StorageRepository class autoloading and automatic connection to the
database (by adding 'self::load();' at the start of each method, which
works with the database).
Add hidden news-item id in news_item page for easier news-item-id
getting in further commits for the comments.