fix news_item.js comment adding
Fix raising TypeError, which appeared, if the user wasn't logged in,
and the comment-form block haven't been placed in the document, so
that the js code haven't been able to add an event listener.
Fix wrong class names, which were adding to the comment item in the
comment list, when the comment got added to the database.