add minor improvements
Change model of the whole database access by making multiple -Repository
classes for each field (news, users, notifications etc.). Each of them
have a link to the original database connection from the StorageRepository,
which now haves only 3 methods: connect(), load() and &getConnection().
Add UserRoles class with user roles constants (ex.: ADMINISTRATOR = 1,
STUDENT = 2 etc.).
Rename all 'date' fields in news field to 'publish_date'.
Add ability for the notification link to be absent: if in the database
some notification have a 'null' value on the link field, then in the
notifications page the link will be hidden (visibility: hidden), because
if empty link is inserted in the anchor (<a></a>) tag, then it leads
to the root.