Fix other-page login json showing
If user was to try to login in news.index or news.show, then
he'll either login successfully, or get an message alert about an
error. But if he was to do that on other page, he'll get to the
page, which showed json response. That wasn't planned.
Problem was that '$' function (jQuery's one) wasn't defined on
other page, because jQuery was included in scripts only on news.show
and news.index pages.
Include jQuery to header, so that it will be available everywhere,
where header is included (in other words - everywhere).